requirements.txt
# Hotel Guest Service Training System - Dependencies
streamlit
requests
openai
chromadb
numpy
PyPDF2
python-docx
pandas
openpyxl
pdfplumber
python-dotenv